refactor(components): Refactor file structure again
