+++ src/components/author/__tests__/author.test.tsx
@@ -1,6 +1,7 @@
+/* eslint-disable @typescript-eslint/explicit-function-return-type */
 import React from 'react';
 import renderer from 'react-test-renderer';
-import Author from '../author';
+import Author from '..';
 
 describe('Author component', () => {
   it('renders correctly', () => {
