+++ src/components/banner/__tests__/banner.test.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/explicit-function-return-type */
 import React from 'react';
 import renderer from 'react-test-renderer';
-import Banner from '../banner';
+import Banner from '..';
 
 describe('Tests for Header component', () => {
   it('renders correctly', () => {
