+++ src/components/banner/index.tsx
