+++ src/components/header/__tests__/header.test.tsx
@@ -1,6 +1,7 @@
+/* eslint-disable @typescript-eslint/explicit-function-return-type */
 import React from 'react';
 import renderer from 'react-test-renderer';
-import Header from '../header';
+import Header from '..';
 
 describe('Tests for Header component', () => {
   it('renders correctly', () => {
