+++ src/components/hero/index.tsx
