+++ src/components/pagination/__tests__/pagination.test.tsx
@@ -1,6 +1,7 @@
+/* eslint-disable @typescript-eslint/explicit-function-return-type */
 import React from 'react';
 import renderer from 'react-test-renderer';
-import Pagination from '../pagination';
+import Pagination from '..';
 import { createPaginationInfo } from '../../../../test/__fixtures__/page';
 
 describe('Pagination component', () => {
