+++ src/containers/AuthorList/index.tsx
@@ -1,6 +1,6 @@
 import { css, SerializedStyles } from '@emotion/core';
 import React from 'react';
-import Author from './author';
+import Author from '../../components/Author';
 
 const list: SerializedStyles = css`
   display: flex;
