+++ src/pages/404.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import Hero from '../components/hero/hero';
-import Layout from '../components/layout/layout';
+import Hero from '../components/Hero';
+import Layout from '../components/Layout';
 
 export default function NotFoundPage(): JSX.Element {
   const title = 'PAGE NOT FOUND';
