+++ src/pages/download.tsx
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
 import { useReleaseHistory } from '../hooks';
-import Layout from '../components/layout/layout';
-import ReleaseTable from '../components/release/release-table';
-import ReleaseToggle from '../components/release/release-toggle';
-import ReleaseCards from '../components/release/release-cards';
+import Layout from '../components/Layout';
+import ReleaseTable from '../components/ReleaseTable';
+import ReleaseToggle from '../components/ReleaseToggle';
+import ReleaseCards from '../components/ReleaseCards';
 
 import { detectOS } from '../util/detectOS';
