+++ src/templates/learn.tsx
@@ -1,13 +1,13 @@
 import { graphql } from 'gatsby';
 import React from 'react';
-import Article from '../components/article/article';
-import Layout from '../components/layout/layout';
-import Navigation from '../components/navigation/navigation';
+import Article from '../components/Article';
+import Layout from '../components/Layout';
+import Navigation from '../containers/Navigation';
 import { LearnPageContext, LearnPageData } from '../types';
 
 import '../styles/article-reader.scss';
 import '../styles/learn.scss';
-import Footer from '../components/footer/footer';
+import Footer from '../components/Footer';
 
 interface Props {
   data: LearnPageData;
